# minimalist-be/Minimalist/Data/Dtos/ReadTodoDto.cs
using System.ComponentModel.DataAnnotations;

namespace Minimalist.Data.Dtos;

public class ReadTodoDto
{
    public string Id { get; set; }
    public string Title { get; set; }
    // public DateTime ReadTime { get; set; } = DateTime.Now;
    public bool IsCompleted { get; set; }
}