# minimalist-fe/src/http/generated/api.schemas.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * TodosAPI
 * OpenAPI spec version: v1
 */
export interface CreateTodoDto {
  /** @minLength 1 */
  title: string;
  isCompleted?: boolean;
}

export interface LoginModel {
  /** @nullable */
  username?: string | null;
  /** @nullable */
  password?: string | null;
}

/**
 * @nullable
 */
export type OperationValue = unknown | null;

export interface Operation {
  operationType?: OperationType;
  /** @nullable */
  path?: string | null;
  /** @nullable */
  op?: string | null;
  /** @nullable */
  from?: string | null;
  /** @nullable */
  value?: OperationValue;
}

export type OperationType = (typeof OperationType)[keyof typeof OperationType];

export const OperationType = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
} as const;

export interface ReadTodoDto {
  /** @minLength 1 */
  id: string;
  /** @nullable */
  title?: string | null;
  isCompleted?: boolean;
}

export type GetTodoParams = {
  skip?: number;
  take?: number;
};