# minimalist-fe/src/todosAPIModel/createTodoDto.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * TodosAPI
 * OpenAPI spec version: v1
 */

export interface CreateTodoDto {
  /** @minLength 1 */
  title: string;
  isCompleted?: boolean;
}