# minimalist-fe/src/todosAPIModel/operation.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * TodosAPI
 * OpenAPI spec version: v1
 */
import type { OperationType } from './operationType';
import type { OperationValue } from './operationValue';

export interface Operation {
  operationType?: OperationType;
  /** @nullable */
  path?: string | null;
  /** @nullable */
  op?: string | null;
  /** @nullable */
  from?: string | null;
  /** @nullable */
  value?: OperationValue;
}