# minimalist-fe/src/todosAPIModel/operationType.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * TodosAPI
 * OpenAPI spec version: v1
 */

export type OperationType = typeof OperationType[keyof typeof OperationType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OperationType = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
} as const;